+++ www/go/core/views/extjs3/usersettings/UserSettingsDialog.js
@@ -109,7 +109,7 @@ go.usersettings.UserSettingsDialog = Ext.extend(go.Window, {
 		this.addPanel(go.usersettings.AccountSettingsPanel);
 		this.addPanel(go.usersettings.LookAndFeelPanel);
 
-		if(go.User.isAdmin) {
+		if(go.Modules.get("core", "core").userRights.mayChangeUsers) {
 			this.addPanel(go.usersettings.VisibleToPanel);
 		}